+++ app/routes/orders/actions/deleteOrder.ts
@@ -1,0 +1,10 @@
+import { type ClientActionFunction } from "@remix-run/react";
+
+import { orderRepository } from "~/repositories/order";
+
+export const deleteOrder: ClientActionFunction = async ({ request }) => {
+  const formData = await request.formData();
+  const id = formData.get("id");
+  await orderRepository.delete(id as string);
+  return null;
+};
